Tests: Move tests for RDA in subfolder
As more tests related to the "analyses" are to be added later.
